move tests into single file
