--- integrations/nvidia/tests/test_nvidia_chat_generator_async.py
@@ -1,191 +1,0 @@
-# SPDX-FileCopyrightText: 2024-present deepset GmbH <info@deepset.ai>
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import os
-from datetime import datetime
-from unittest.mock import AsyncMock, patch
-
-import pytest
-import pytz
-from haystack.dataclasses import ChatMessage, StreamingChunk
-from haystack.tools import Tool
-from openai import AsyncOpenAI
-from openai.types.chat import ChatCompletion, ChatCompletionMessage
-from openai.types.chat.chat_completion import Choice
-
-from haystack_integrations.components.generators.nvidia.chat.chat_generator import NvidiaChatGenerator
-
-
-@pytest.fixture
-def chat_messages():
-    return [
-        ChatMessage.from_system("You are a helpful assistant"),
-        ChatMessage.from_user("What's the capital of France"),
-    ]
-
-
-def weather(city: str):
-    """Get weather for a given city."""
-    return f"The weather in {city} is sunny and 32°C"
-
-
-@pytest.fixture
-def tools():
-    tool_parameters = {
-        "type": "object",
-        "properties": {"city": {"type": "string"}},
-        "required": ["city"],
-    }
-    tool = Tool(
-        name="weather",
-        description="useful to determine the weather in a given location",
-        parameters=tool_parameters,
-        function=weather,
-    )
-
-    return [tool]
-
-
-@pytest.fixture
-def mock_async_chat_completion():
-    """
-    Mock the Async OpenAI API completion response and reuse it for async tests
-    """
-    with patch(
-        "openai.resources.chat.completions.AsyncCompletions.create",
-        new_callable=AsyncMock,
-    ) as mock_chat_completion_create:
-        completion = ChatCompletion(
-            id="foo",
-            model="meta/llama-3.1-8b-instruct",
-            object="chat.completion",
-            choices=[
-                Choice(
-                    finish_reason="stop",
-                    logprobs=None,
-                    index=0,
-                    message=ChatCompletionMessage(content="Hello world!", role="assistant"),
-                )
-            ],
-            created=int(datetime.now(tz=pytz.timezone("UTC")).timestamp()),
-            usage={
-                "prompt_tokens": 57,
-                "completion_tokens": 40,
-                "total_tokens": 97,
-            },
-        )
-        # For async mocks, the return value should be awaitable
-        mock_chat_completion_create.return_value = completion
-        yield mock_chat_completion_create
-
-
-class TestNvidiaChatGeneratorAsync:
-    def test_init_default_async(self, monkeypatch):
-        monkeypatch.setenv("NVIDIA_API_KEY", "test-api-key")
-        component = NvidiaChatGenerator()
-
-        assert isinstance(component.async_client, AsyncOpenAI)
-        assert component.async_client.api_key == "test-api-key"
-        assert not component.generation_kwargs
-
-    @pytest.mark.asyncio
-    async def test_run_async(self, chat_messages, mock_async_chat_completion, monkeypatch):  # noqa: ARG002
-        monkeypatch.setenv("NVIDIA_API_KEY", "fake-api-key")
-        component = NvidiaChatGenerator()
-        response = await component.run_async(chat_messages)
-
-        # check that the component returns the correct ChatMessage response
-        assert isinstance(response, dict)
-        assert "replies" in response
-        assert isinstance(response["replies"], list)
-        assert len(response["replies"]) == 1
-        assert [isinstance(reply, ChatMessage) for reply in response["replies"]]
-
-    @pytest.mark.asyncio
-    async def test_run_async_with_params(self, chat_messages, mock_async_chat_completion, monkeypatch):
-        monkeypatch.setenv("NVIDIA_API_KEY", "fake-api-key")
-        component = NvidiaChatGenerator(generation_kwargs={"max_tokens": 10, "temperature": 0.5})
-        response = await component.run_async(chat_messages)
-
-        # check that the component calls the OpenAI API with the correct parameters
-        _, kwargs = mock_async_chat_completion.call_args
-        assert kwargs["max_tokens"] == 10
-        assert kwargs["temperature"] == 0.5
-
-        # check that the component returns the correct response
-        assert isinstance(response, dict)
-        assert "replies" in response
-        assert isinstance(response["replies"], list)
-        assert len(response["replies"]) == 1
-        assert [isinstance(reply, ChatMessage) for reply in response["replies"]]
-
-    @pytest.mark.asyncio
-    async def test_run_async_with_extra_body(self, chat_messages, mock_async_chat_completion, monkeypatch):
-        monkeypatch.setenv("NVIDIA_API_KEY", "fake-api-key")
-        extra_body = {
-            "guardrails": {"config_id": "demo-self-check-input-output"},
-        }
-        component = NvidiaChatGenerator(generation_kwargs={"extra_body": extra_body})
-        response = await component.run_async(chat_messages)
-
-        # check that the component calls the OpenAI API with the correct parameters
-        _, kwargs = mock_async_chat_completion.call_args
-        assert kwargs["extra_body"] == extra_body
-        assert kwargs["model"] == "meta/llama-3.1-8b-instruct"
-        assert kwargs["messages"] == [
-            {"role": "system", "content": "You are a helpful assistant"},
-            {"role": "user", "content": "What's the capital of France"},
-        ]
-
-        # check that the component returns the correct response
-        assert isinstance(response, dict)
-        assert "replies" in response
-        assert isinstance(response["replies"], list)
-        assert len(response["replies"]) == 1
-        assert [isinstance(reply, ChatMessage) for reply in response["replies"]]
-
-    @pytest.mark.skipif(
-        not os.environ.get("NVIDIA_API_KEY", None),
-        reason="Export an env var called NVIDIA_API_KEY containing the NVIDIA API key to run this test.",
-    )
-    @pytest.mark.integration
-    @pytest.mark.asyncio
-    async def test_live_run_async(self):
-        chat_messages = [ChatMessage.from_user("What's the capital of France")]
-        component = NvidiaChatGenerator()
-        results = await component.run_async(chat_messages)
-        assert len(results["replies"]) == 1
-        message: ChatMessage = results["replies"][0]
-        assert "Paris" in message.text
-        assert "meta/llama-3.1-8b-instruct" in message.meta["model"]
-        assert message.meta["finish_reason"] == "stop"
-
-    @pytest.mark.skipif(
-        not os.environ.get("NVIDIA_API_KEY", None),
-        reason="Export an env var called NVIDIA_API_KEY containing the NVIDIA API key to run this test.",
-    )
-    @pytest.mark.integration
-    @pytest.mark.asyncio
-    async def test_live_run_streaming_async(self):
-        counter = 0
-        responses = ""
-
-        async def callback(chunk: StreamingChunk):
-            nonlocal counter
-            nonlocal responses
-            counter += 1
-            responses += chunk.content if chunk.content else ""
-
-        component = NvidiaChatGenerator(streaming_callback=callback)
-        results = await component.run_async([ChatMessage.from_user("What's the capital of France?")])
-
-        assert len(results["replies"]) == 1
-        message: ChatMessage = results["replies"][0]
-        assert "Paris" in message.text
-
-        assert "meta/llama-3.1-8b-instruct" in message.meta["model"]
-        assert message.meta["finish_reason"] == "stop"
-
-        assert counter > 1
-        assert "Paris" in responses
